Add overrides for all SysEngRun2 runs.
Until updates to discovery are made to allow program dates and CEUs to be passed through to the template, we are stuck with this. They should be in place by the next round. We would also pass through the logo to the template which would remove all of these overrides.