Allow override of program title.
Titles longer than 42 chars are split into title and subtitle for
formatting purposes. Currently subtitle is not passed into the template
content, so for the time being program titles must be overriden in
order for the certificate to display the correct title.

This will be fixed once WL-1120 is complete and subtitle can be added
to the context.