chore: remove glob dependency (#582)
I was looking at a renovate PR that is trying to update `glob` from v8 --> v9. After taking another look, I realized that we aren't even using `glob` anymore. We removed its use in 2019. This PR removes `glob` from our `package.json` and removes the only use of the package in this repo.